Added money, moved initialization of enemy sprites
Added a money attribute to the Player class, and a method to add money to the player when they kill an enemy.

Moved the initialization of enemy sprites from the class's __init__() to the main stack frame of the file containing the Enemy class. This was done to improve efficiency.