Implemented buying of ammo
Implemented ability to buy ammo with player's money, added a buy platform where the player can hit "B" to purchase more ammo. Implemented through methods check_buy() and buy_ammo() in the Player class. check_buy() assesses if the player is standing on the buy platform as well as if they have an appropriate amount of money to buy ammo, and displays an appropriate message with instructions. If the player hits "B", buy_ammo() is called which adds the ammo to the player's current and total ammo.