Added a 2nd enemy, fixed bullet bug
Added a second enemy, the Bat as a subclass of the Enemy class. This enemy has different movement where it follows the characters location inaccurately but can move very fast.

Fixed a bug where bullets would cause the game to crash if they struck two enemies at once. Figured out that a for loop that looped over enemies would remove a bullet, and then try to remove the bullet from the bullet list again when it looped to the other enemy, causing an error.