Enemy class back in main, added Mushroom enemy
Moved the Enemy abstract class as well as it's subclasses back into main.py in order to implement better efficiency with loading in enemy sprites every time an enemy is made.

Added a Mushroom enemy which moves towards the player and once close enough, stops and primes before exploding and damaging the player if they are near.