Added in Wave class
Added in the Wave class, which handles the waves in the game. Each wave spawns in a certain amount of enemies and handles these enemies election of which enemies to spawn, their movement, the damage they deal, damage dealt to them, and their death. This class also checks the status of the current wave, and prompts the player to start the wave, and informs the player when the wave has ended.