Updated how timer sound is handled, fixed control instruction bug
Updated how the timer sound when the mushroom is about to explode is handled. Made it into an attribute of the Mushroom enemy class, allows for the sound to be stopped for a specific enemy instead of all of them if multiple Mushroom enemies are priming to explode and playing the timer sound.

Fixed a bug where the control instructions display after every round, made it so that they only display when the current wave is 0.